# backend/endoscope/endoscope-api/target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: endoscope@localhost
    name: endoscope
  description: API for endoscope
  title: API for endoscope
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /images:
    post:
      description: Edit description of Image
      operationId: editImageDescription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageDescriptionUpsert'
        description: Image description to edit
        required: true
      responses:
        "200":
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal Server Error
      summary: Edit description of Image
      tags:
      - image
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: image
  /images/{imageId}:
    delete:
      description: Delete image
      operationId: deleteImage
      parameters:
      - explode: false
        in: path
        name: imageId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: Successful operation
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal Server Error
      summary: Delete image
      tags:
      - image
      x-accepts: application/json
      x-tags:
      - tag: image
    get:
      description: Get image  by ID
      operationId: getImageById
      parameters:
      - explode: false
        in: path
        name: imageId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageEntity'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal Server Error
      summary: Get image  by ID
      tags:
      - image
      x-accepts: application/json
      x-tags:
      - tag: image
  /images/directory/{directoryId}:
    get:
      description: Get Images by directory ID
      operationId: getImagesByDirectoryId
      parameters:
      - explode: false
        in: path
        name: directoryId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ImageEntity'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal Server Error
      summary: Get Images by directory ID
      tags:
      - image
      x-accepts: application/json
      x-tags:
      - tag: image
    put:
      description: Create image in directory
      operationId: createImageInDirectory
      parameters:
      - explode: false
        in: path
        name: directoryId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/ImageEntity'
              type: array
        description: Image to create
        required: true
      responses:
        "200":
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal Server Error
      summary: Create image in directory
      tags:
      - image
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: image
components:
  schemas:
...
    ImageEntity:
      example:
        imageId: 0
        imageName: imageName
        uploadDate: ""
        imageData: imageData
        description: description
        directory: 6
        uploadedBy: uploadedBy
      properties:
        imageId:
          description: The image ID
          type: integer
        imageName:
          description: The image name
          type: string
        directory:
          description: The directory ID where the image belongs
          type: integer
        uploadedBy:
          description: The name of the user who uploaded the image
          type: string
        uploadDate:
          description: The upload date of the image in milliseconds since Unix epoch
          type: long
        description:
          description: Description of the image
          type: string
        imageData:
          description: Binary image data in base64 format
          format: byte
          type: string
      type: object
    ImageDescriptionUpsert:
      example:
        imageId: ""
        description: description
      properties:
        imageId:
          description: The image ID
          type: long
        description:
          description: The image description
          type: string
      type: object
    InternalServerError:
      properties:
        message:
          description: The error message
          type: string
        code:
          description: The error code
          type: integer
      type: object